src/api_proxy/requirements.txt
# API Proxy Dependencies

# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
email-validator>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# HTTP Client
httpx>=0.25.2
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
